+++ lib/apiScenario/apiScenarioSchema.ts
@@ -114,7 +114,10 @@ export const ApiScenarioDefinition: Schema & {
           $ref: "#/definitions/StepRestOperation",
         },
         {
-          $ref: "#/definitions/StepArmTemplateDeployment",
+          $ref: "#/definitions/StepArmTemplate",
+        },
+        {
+          $ref: "#/definitions/StepArmDeploymentScript",
         },
         {
           $ref: "#/definitions/StepRawCall",
@@ -230,7 +233,7 @@ export const ApiScenarioDefinition: Schema & {
       },
       required: ["operationId", "resourceName"],
     },
-    StepArmTemplateDeployment: {
+    StepArmTemplate: {
       type: "object",
       allOf: [
         {
@@ -238,11 +241,43 @@ export const ApiScenarioDefinition: Schema & {
         },
       ],
       properties: {
-        armTemplateDeployment: {
+        armTemplate: {
           type: "string",
         },
       },
-      required: ["armTemplateDeployment"],
+      required: ["armTemplate"],
+    },
+    StepArmDeploymentScript: {
+      type: "object",
+      allOf: [
+        {
+          $ref: "#/definitions/StepBase",
+        },
+      ],
+      properties: {
+        armDeploymentScript: {
+          type: "string",
+        },
+        arguments: {
+          type: "string",
+        },
+        environmentVariables: {
+          type: "array",
+          items: {
+            type: "object",
+            properties: {
+              name: {
+                type: "string",
+              },
+              value: {
+                type: "string",
+              },
+            },
+            required: ["name", "value"],
+          },
+        },
+      },
+      required: ["armDeploymentScript"],
     },
     StepRawCall: {
       type: "object",